Update ETL3 to use new shared components
