--- migration_steps/acquire_target_ids/app/mapping_definitions/client_addresses_mapping.json
@@ -1,107 +1,0 @@
-{
-    "id": {
-        "casrec_table": "",
-        "casrec_column_name": "",
-        "alias": "",
-        "requires_transformation": "",
-        "default_value": "",
-        "calculated": "",
-        "is_pk": true,
-        "fk_parents": "",
-        "is_complete": true
-    },
-    "person_id": {
-        "casrec_table": "",
-        "casrec_column_name": "",
-        "alias": "",
-        "requires_transformation": "",
-        "default_value": "",
-        "calculated": "",
-        "is_pk": "",
-        "fk_parents": "persons:id",
-        "is_complete": true
-    },
-    "address_lines": {
-        "casrec_table": "PAT",
-        "casrec_column_name": [
-            "Adrs1",
-            "Adrs2"
-        ],
-        "alias": [
-            "Adrs1",
-            "Adrs2"
-        ],
-        "requires_transformation": "squash_columns",
-        "default_value": "",
-        "calculated": "",
-        "is_pk": "",
-        "fk_parents": "",
-        "is_complete": true
-    },
-    "town": {
-        "casrec_table": "PAT",
-        "casrec_column_name": "Adrs3",
-        "alias": "Adrs3",
-        "requires_transformation": "",
-        "default_value": "",
-        "calculated": "",
-        "is_pk": "",
-        "fk_parents": "",
-        "is_complete": true
-    },
-    "county": {
-        "casrec_table": "PAT",
-        "casrec_column_name": "Adrs4",
-        "alias": "Adrs4",
-        "requires_transformation": "",
-        "default_value": "",
-        "calculated": "",
-        "is_pk": "",
-        "fk_parents": "",
-        "is_complete": true
-    },
-    "postcode": {
-        "casrec_table": "PAT",
-        "casrec_column_name": "Postcode",
-        "alias": "Postcode",
-        "requires_transformation": "",
-        "default_value": "",
-        "calculated": "",
-        "is_pk": "",
-        "fk_parents": "",
-        "is_complete": true
-    },
-    "country": {
-        "casrec_table": "PAT",
-        "casrec_column_name": "Adrs5",
-        "alias": "Adrs5",
-        "requires_transformation": "",
-        "default_value": "",
-        "calculated": "",
-        "is_pk": "",
-        "fk_parents": "",
-        "is_complete": true
-    },
-    "type": {
-        "casrec_table": "",
-        "casrec_column_name": "",
-        "alias": "",
-        "requires_transformation": "",
-        "default_value": "Primary",
-        "calculated": "",
-        "is_pk": "",
-        "fk_parents": "",
-        "is_complete": true
-    },
-    "isairmailrequired": {
-        "casrec_table": "PAT",
-        "casrec_column_name": "Foreign",
-        "alias": "Foreign",
-        "requires_transformation": "convert_to_bool",
-        "default_value": "",
-        "calculated": "",
-        "is_pk": "",
-        "fk_parents": "",
-        "is_complete": true
-    }
-}
