--- migration_steps/acquire_target_ids/app/mapping_definitions/meta/mapping_progress_summary.json
@@ -1,34 +1,0 @@
-{
-    "worksheets": {
-        "client_persons": {
-            "total_rows": 31,
-            "total_unmapped": 5,
-            "total_mapped": 26,
-            "percentage_complete": 84
-        },
-        "client_addresses": {
-            "total_rows": 9,
-            "total_unmapped": 0,
-            "total_mapped": 9,
-            "percentage_complete": 100
-        },
-        "client_phonenumbers": {
-            "total_rows": 6,
-            "total_unmapped": 0,
-            "total_mapped": 6,
-            "percentage_complete": 100
-        }
-    },
-    "total": {
-        "worksheets": {
-            "total_sheets": 3,
-            "total_complete": 2
-        },
-        "fields": {
-            "total_fields": 46,
-            "total_unmapped": 5,
-            "total_mapped": 41,
-            "percentage_complete": 89
-        }
-    }
-}
