+++ migration_steps/shared/db_helpers.py
@@ -1,9 +1,24 @@
 import os
+import psycopg2
+import pandas as pd
+import numpy as np
 
-def execute_sql_file(sql_path, filename, conn):
+from psycopg2.extensions import register_adapter, AsIs
+psycopg2.extensions.register_adapter(np.int64, psycopg2._psycopg.AsIs)
+
+
+def execute_sql_file(sql_path, filename, conn, schema='public'):
     cursor = conn.cursor()
     sql_file = open(sql_path / filename, 'r')
-    cursor.execute(sql_file.read())
+
+    try:
+        cursor.execute(sql_file.read().replace('{schema}', str(schema)))
+        conn.commit()
+    except (Exception, psycopg2.DatabaseError) as error:
+        print("Error: %s" % error)
+        conn.rollback()
+        cursor.close()
+        return 1
     cursor.close()
 
 
@@ -31,3 +46,29 @@ def result_from_sql_file(sql_path, filename, conn):
     result = cursor.fetchone()[0]
     cursor.close()
     return result
+
+
+def df_from_sql_file(sql_path, filename, conn, schema="public"):
+    sql_file = open(sql_path / filename, 'r')
+    sql = sql_file.read().replace('{schema}', str(schema))
+    return pd.read_sql_query(sql, con=conn, index_col=None)
+
+
+def execute_insert(conn, df, table):
+    tuples = [tuple(x) for x in df.to_numpy()]
+    cols = ','.join(list(df.columns))
+
+    cursor = conn.cursor()
+    row_str_template = ','.join(['%s']*len(df.columns))
+    values = [cursor.mogrify('('+row_str_template+')', tup).decode('utf8') for tup in tuples]
+    query = "INSERT INTO %s(%s) VALUES " % (table, cols) + ",".join(values)
+
+    try:
+        cursor.execute(query, tuples)
+        conn.commit()
+    except (Exception, psycopg2.DatabaseError) as error:
+        print("Error: %s" % error)
+        conn.rollback()
+        cursor.close()
+        return 1
+    cursor.close()
